requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
